Renamed the GitHub repository
Note that I'm currently working on a Go (golang) implementation of
Denada.  I am not sure how much use this version of Denada will see
in the immediate future.  But I'd like to keep the two of them in
step and perhaps publish a formal Denada spec at some point.